+++ core/resolve.ts
@@ -3,7 +3,9 @@ import {
   distinctBy,
   head,
   init,
+  isEmptyObject,
   isFunction,
+  isLength0,
   isRegExp,
   isString,
   isUndefined,
@@ -12,25 +14,24 @@ import {
   propPath,
   Root,
   Rule,
-  sortBy,
   tail,
 } from "../deps.ts";
 import { astify } from "./ast.ts";
 import { isCSSDefinition, isCSSObject } from "./utils/assert.ts";
-import { escapeRegExp } from "./utils/escape.ts";
 import type {
-  Config,
   CSSObject,
   ModifierContext,
   ModifierMap,
-  PostProcessor,
+  PreProcessor,
+  Preset,
   Specifier,
   SpecifierContext,
   SpecifierHandler,
   SpecifierMap,
+  StaticConfig,
+  StaticContext,
   Syntax,
   Theme,
-  ThemeContext,
 } from "./types.ts";
 
 function firstSplit(
@@ -65,16 +66,14 @@ class MockRegExpExecArray extends Array<string> {
   }
 }
 
-export function resolveMappedSpecifier(
+export function resolveDeepMapSpecifier(
   value: string | string[],
-  mappedSpecifier: MappedSpecifier,
+  deepMapSpecifier: DeepMapSpecifier,
   context:
-    & Omit<SpecifierContext, "className" | "key" | "parentKey" | "path">
+    & Omit<SpecifierContext, "key" | "parentKey" | "path">
     & { parentKey?: string; key?: string },
 ): Root | undefined {
-  const className = `.${escapeRegExp(context.token)}`;
   const paths = leftSplit(value, context.separator);
-
   for (const path of paths) {
     const first = head(path) ?? "DEFAULT";
     const rest = tail(path);
@@ -82,27 +81,27 @@ export function resolveMappedSpecifier(
     const specifierContext: SpecifierContext = {
       ...context,
       parentKey: context.key,
-      className,
       key: first,
       path,
     };
-    const has = mappedSpecifier.has(first);
+    const has = deepMapSpecifier.has(first);
     if (has) {
-      const definition = mappedSpecifier.get(first)!;
+      const definition = deepMapSpecifier.get(first)!;
       if (isFunction(definition)) {
+        if (!isLength0(rest)) continue;
         const result = definition(new MockRegExpExecArray(), specifierContext);
-        specifierContext.parentKey = first;
         if (isUndefined(result)) continue;
-        return handleCSSObject(result, className);
+        return handleCSSObject(result, specifierContext.className);
       }
 
       if (definition instanceof Map) {
-        return resolveMappedSpecifier(rest, definition, specifierContext);
+        return resolveDeepMapSpecifier(rest, definition, specifierContext);
       }
-      return handleCSSObject(definition, className);
+      if (isEmptyObject(definition)) return;
+      return handleCSSObject(definition, specifierContext.className);
     }
 
-    for (const [key, m] of mappedSpecifier) {
+    for (const [key, m] of deepMapSpecifier) {
       if (isRegExp(key)) {
         const regExpExecResult = key.exec(first);
         if (!regExpExecResult) continue;
@@ -111,12 +110,13 @@ export function resolveMappedSpecifier(
           const result = m(regExpExecResult, specifierContext);
           if (isUndefined(result)) continue;
 
-          return handleCSSObject(result, className);
+          return handleCSSObject(result, specifierContext.className);
         }
         if (m instanceof Map) {
-          return resolveMappedSpecifier(rest, m, specifierContext);
+          return resolveDeepMapSpecifier(rest, m, specifierContext);
         }
-        return handleCSSObject(m, className);
+        if (isEmptyObject(m)) return;
+        return handleCSSObject(m, specifierContext.className);
       }
     }
   }
@@ -126,10 +126,10 @@ function handleCSSObject(cssObject: CSSObject, selector: string): Root {
   if (cssObject instanceof Root) {
     return cssObject;
   } else if (isCSSDefinition(cssObject)) {
-    return new Root({ nodes: astify(cssObject.value) });
+    return astify(cssObject.value);
   } else {
     return new Root({
-      nodes: [new Rule({ selector, nodes: astify(cssObject) })],
+      nodes: [new Rule({ selector, nodes: astify(cssObject).nodes })],
     });
   }
 }
@@ -138,7 +138,7 @@ function handleCSSObject(cssObject: CSSObject, selector: string): Root {
 export function resolveTheme(
   identifier: string,
   themeRoot: string,
-  { separator, theme }: ThemeContext,
+  { separator, theme }: Pick<StaticContext, "theme" | "separator">,
 ): string | undefined {
   const paths = leftSplit(identifier, separator);
   for (const path of paths) {
@@ -153,7 +153,7 @@ export function resolveTheme(
 export function $resolveTheme(
   identifier: string,
   themeRoot: string,
-  { separator, theme }: ThemeContext,
+  { separator, theme }: Pick<StaticContext, "theme" | "separator">,
 ): Theme | string | undefined {
   const recursive = (
     path: string[],
@@ -181,46 +181,60 @@ function pickByName({ name }: { name: string }): string {
   return name;
 }
 
-export function resolvePostProcessor(
-  ...postProcessors: PostProcessor[]
-): PostProcessor[] {
-  const processor = distinctBy(postProcessors, pickByName);
-  return sortBy(processor, ({ order }) => order ?? 0);
+export function resolvePreProcessor(
+  ...postProcessors: PreProcessor[]
+): PreProcessor[] {
+  return distinctBy(postProcessors, pickByName);
 }
 
 export function resolveSyntax(...syntaxes: Syntax[]): Syntax[] {
   return distinctBy(syntaxes, pickByName);
 }
 
+export function resolvePreset(
+  preset: Readonly<Readonly<Preset>[]>,
+  context: Readonly<Omit<StaticContext, "theme">>,
+): Omit<StaticConfig, "preset">[] {
+  return distinctBy(preset, pickByName).map(({ fn }) => {
+    const {
+      syntax = [],
+      specifierMap = {},
+      modifierMap = {},
+      theme = {},
+      preProcess = [],
+      css = {},
+    } = fn(context);
+    return {
+      syntax,
+      specifierMap,
+      modifierMap,
+      theme,
+      preProcess,
+      css,
+    };
+  });
+}
+
 /** resolve config to deep merge */
 export function resolveConfig(
   {
-    syntaxes: _syntaxes = [],
-    presets = [],
+    syntax: _syntax = [],
+    preset = [],
     specifierMap: _specifierMap = {},
     theme: _theme = {},
-    postProcess: _postProcess = [],
+    preProcess: _postProcess = [],
     modifierMap: _modifierMap = {},
+    css: _css = {},
   }: Readonly<
     Partial<
-      Pick<
-        Config,
-        | "specifierMap"
-        | "modifierMap"
-        | "theme"
-        | "presets"
-        | "syntaxes"
-        | "postProcess"
-      >
+      StaticConfig
     >
   >,
+  context: Readonly<Omit<StaticContext, "theme">>,
 ):
-  & Pick<
-    Config,
-    "modifierMap" | "theme" | "syntaxes" | "postProcess"
-  >
-  & { mappedSpecifier: MappedSpecifier } {
-  const _presets = distinctBy(presets, pickByName);
+  & Omit<StaticConfig, "specifierMap" | "preset">
+  & { deepMapSpecifier: DeepMapSpecifier } {
+  const _presets = resolvePreset(preset, context);
   const modifierMap = _presets.map(({ modifierMap }) => modifierMap)
     .reduce((acc, cur) => {
       return deepMerge(acc, cur);
@@ -231,35 +245,40 @@ export function resolveConfig(
     },
     _theme,
   );
-  const syntaxes = resolveSyntax(
-    ..._syntaxes,
-    ..._presets.map(({ syntaxes }) => syntaxes).flat(),
+  const syntax = resolveSyntax(
+    ..._syntax,
+    ..._presets.map(({ syntax }) => syntax).flat(),
   );
-  const mappedSpecifier = mergeSpecifierMap(
-    _presets.map(({ specifierMap }) => specifierMap),
+  const deepMapSpecifier = mergeSpecifierMap(
+    [..._presets.map(({ specifierMap }) => specifierMap), _specifierMap],
   );
 
-  const postProcess = resolvePostProcessor(
+  const preProcess = resolvePreProcessor(
     ..._postProcess,
-    ..._presets.map(({ postProcessor }) => postProcessor).flat(),
+    ..._presets.map(({ preProcess }) => preProcess).flat(),
+  );
+  const css = [..._presets.map(({ css }) => css), _css].reduce(
+    (acc, cur) => deepMerge(acc, cur),
+    {},
   );
 
   return {
-    mappedSpecifier,
+    deepMapSpecifier,
     theme,
     modifierMap,
-    syntaxes,
-    postProcess,
+    syntax,
+    preProcess,
+    css,
   };
 }
 
 type TreeMap<Leaf, P> = Map<P, Leaf | TreeMap<Leaf, P>>;
 
-type MappedSpecifier = TreeMap<SpecifierHandler | CSSObject, string | RegExp>;
+type DeepMapSpecifier = TreeMap<SpecifierHandler | CSSObject, string | RegExp>;
 
 export function mergeSpecifierMap(
   specifierMaps: SpecifierMap[],
-): MappedSpecifier {
+): DeepMapSpecifier {
   const recursive = (
     m: Specifier,
     map: Map<string | RegExp, any>,