+++ preset_tw/mod.ts
@@ -1,4 +1,4 @@
-export * from "./plugin.ts";
+export { type Option as TwOption, presetTw } from "./plugin.ts";
 export * from "./specifier/mod.ts";
 export * from "./theme/mod.ts";
 export * from "./modifier/mod.ts";