+++ preset_typography/mod.ts
@@ -1,2 +1,2 @@
 export { plugin as presetTypography } from "./plugin.ts";
-export type { PresetOptions as PresetTypographyOptions } from "./types.ts";
+export type { PresetOptions as TypographyOptions } from "./types.ts";