+++ preset_typography/specifier/prose_test.ts
@@ -4,10 +4,9 @@ import {
   toAst,
   transformSelector,
 } from "./prose.ts";
-import { Root } from "../../deps.ts";
 import { astify } from "../../core/ast.ts";
 import { expect, type ParamReturn, test } from "../../dev_deps.ts";
-import type { Tree } from "../../core/types.ts";
+import type { BinaryTree } from "../../core/types.ts";
 
 test("transformWhere", () => {
   const table: [...ParamReturn<typeof transformSelector>][] = [
@@ -118,13 +117,16 @@ test("isolateEntries", () => {
 });
 
 test("toAst", () => {
-  const table: [...Parameters<typeof toAst>, Tree<string | number>][] = [
+  const table: [...Parameters<typeof toAst>, BinaryTree<string | number>][] = [
     [{}, {}, {}],
     [{ a: { color: "red" } }, { a: false }, {}],
-    [{ a: { fontWeight: 600 } }, { a: { fontWeight: false } }, { a: {} }],
-    [{ a: { fontWeight: 600 } }, { a: { "font-weight": false } }, { a: {} }],
-    [{ a: { "font-weight": 600 } }, { a: { fontWeight: false } }, { a: {} }],
-    [{ a: { "font-weight": 600 } }, { a: { "font-weight": false } }, { a: {} }],
+    [{ a: { fontWeight: 600 } }, { a: { fontWeight: false } }, {}],
+    [{ a: {} }, {}, { a: {} }],
+    [{ a: {}, b: {} }, { a: false }, { b: {} }],
+    [{ a: {}, b: {} }, { a: false, b: { empty: false } }, { b: {} }],
+    [{ a: { fontWeight: 600 } }, { a: { "font-weight": false } }, {}],
+    [{ a: { "font-weight": 600 } }, { a: { fontWeight: false } }, {}],
+    [{ a: { "font-weight": 600 } }, { a: { "font-weight": false } }, {}],
     [{
       "a:not(h1,h2 ,h3,   h4)": {
         color: "red",
@@ -143,9 +145,7 @@ test("toAst", () => {
           color: false,
         },
       },
-      {
-        a: {},
-      },
+      {},
     ],
     [
       {
@@ -277,7 +277,7 @@ test("toAst", () => {
           color: false,
         },
       },
-      { "h1:hover": {} },
+      {},
     ],
     [
       {
@@ -327,40 +327,36 @@ test("toAst", () => {
         },
       },
       {
-        ":not(pre) > code::before": {},
         ":not(pre) > code::after": {
           content: '"`"',
         },
       },
     ],
-    // OK but can't test it because distribute object.
-    // [
-    //   {
-    //     "h1, h2": {
-    //       color: "red",
-    //     },
-    //     h1: {
-    //       fontWeight: 600,
-    //     },
-    //   },
-    //   {
-    //     "h1, h2": {
-    //       color: false,
-    //     },
-    //   },
-    //   {
-    //     ...{ h1: {} },
-    //     "h1": {
-    //       fontWeight: 600,
-    //     },
-    //     h2: {},
-    //   },
-    // ],
+    [
+      {
+        "h1, h2": {
+          color: "red",
+        },
+        h1: {
+          fontWeight: 600,
+        },
+      },
+      {
+        "h1, h2": {
+          color: false,
+        },
+      },
+      {
+        "h1": {
+          fontWeight: 600,
+        },
+      },
+    ],
   ];
 
   table.forEach(([css, disableMap, result]) =>
     expect(toAst(css, disableMap).toString()).toEqual(
-      new Root({ nodes: astify(result) }).toString(),
+      astify(result).toString(),
     )
   );
 });